requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.1
itsdangerous==0.24
wsgiref==0.1.2
RPi.GPIO==0.5.11
picamera==1.9
